Begins splitting webpack configurations
 - Each major config will have it's own module that defines
    their configurations. More splitting can be done within that
    to help with readability and highlight the more customizable
    parts, based on the project's needs.
